Extract Interpreter Debuging into dedicated interpreter.debug plugin
This avoids allows to remove the dependency of henshin.interpreter on
eclipse.debug.core.
