+++ plugins/org.eclipse.emf.henshin.interpreter.debug/META-INF/MANIFEST.MF
@@ -1,0 +1,15 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: %pluginName
+Bundle-Vendor: %providerName
+Bundle-SymbolicName: org.eclipse.emf.henshin.interpreter.debug;singleton:=true
+Bundle-Version: 1.9.0.qualifier
+Automatic-Module-Name: org.eclipse.emf.henshin.interpreter.debug
+Bundle-Localization: plugin
+Bundle-RequiredExecutionEnvironment: JavaSE-11
+Export-Package: org.eclipse.emf.henshin.interpreter.debug,
+ org.eclipse.emf.henshin.interpreter.impl.debug,
+ org.eclipse.emf.henshin.interpreter.matching.conditions.debug
+Require-Bundle: org.eclipse.emf.henshin.interpreter;bundle-version="1.9.0";visibility:=reexport,
+ org.eclipse.debug.core;bundle-version="3.13.0",
+ org.eclipse.core.runtime;bundle-version="3.15.0"
