+++ plugins/org.eclipse.emf.henshin.interpreter.debug/build.properties
@@ -1,0 +1,13 @@
+# <copyright>
+# Copyright (c) 2025-2025 Henshin developers. All rights reserved. 
+# This program and the accompanying materials are made available 
+# under the terms of the Eclipse Public License v1.0 which 
+# accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-v10.html
+# </copyright>
+
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .,\
+               plugin.properties
