+++ plugins/org.eclipse.emf.henshin.interpreter.debug/plugin.properties
@@ -1,0 +1,10 @@
+# <copyright>
+# Copyright (c) 2010-2014 Henshin developers. All rights reserved.
+# This program and the accompanying materials are made available
+# under the terms of the Eclipse Public License v1.0 which
+# accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-v10.html
+# </copyright>
+
+pluginName=Henshin Interpreter Debug
+providerName=Eclipse Modeling Project
