+++ plugins/org.eclipse.emf.henshin.interpreter.ui/META-INF/MANIFEST.MF
@@ -14,7 +14,7 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.6.0",
  org.eclipse.ui.ide;bundle-version="3.6.0",
  org.eclipse.compare;bundle-version="3.5.0",
  org.eclipse.emf.henshin.diagram;bundle-version="0.9.0",
- org.eclipse.emf.henshin.interpreter;bundle-version="0.9.0",
+ org.eclipse.emf.henshin.interpreter.debug;bundle-version="1.9.0",
  org.eclipse.emf.compare;bundle-version="1.1.0",
  org.eclipse.emf.common.ui;bundle-version="2.6.0",
  org.eclipse.emf.ecore.xmi;bundle-version="2.5.0",
