+++ plugins/org.eclipse.emf.henshin.interpreter.ui/src/org/eclipse/emf/henshin/interpreter/ui/debug/DebugModelPresentation.java
@@ -3,7 +3,7 @@ package org.eclipse.emf.henshin.interpreter.ui.debug;
 import org.eclipse.debug.core.model.IValue;
 import org.eclipse.debug.ui.IDebugModelPresentation;
 import org.eclipse.debug.ui.IValueDetailListener;
-import org.eclipse.emf.henshin.interpreter.matching.conditions.HenshinBreakpoint;
+import org.eclipse.emf.henshin.interpreter.matching.conditions.debug.HenshinBreakpoint;
 import org.eclipse.jface.viewers.LabelProvider;
 import org.eclipse.ui.IEditorInput;
