+++ plugins/org.eclipse.emf.henshin.interpreter/META-INF/MANIFEST.MF
@@ -10,12 +10,9 @@ Bundle-ActivationPolicy: lazy
 Require-Bundle: org.eclipse.emf.ecore.xmi;bundle-version="2.5.0",
  org.eclipse.emf.ecore.change;bundle-version="2.5.0";visibility:=reexport,
  org.eclipse.emf.henshin.model;bundle-version="0.9.0";visibility:=reexport,
- org.eclipse.core.runtime;bundle-version="3.6.0",
- org.eclipse.debug.core,
  org.eclipse.emf.common,
  org.eclipse.emf.ecore
 Export-Package: org.eclipse.emf.henshin.interpreter,
- org.eclipse.emf.henshin.interpreter.debug,
  org.eclipse.emf.henshin.interpreter.impl,
  org.eclipse.emf.henshin.interpreter.info,
  org.eclipse.emf.henshin.interpreter.matching.conditions,
