+++ plugins/org.eclipse.emf.henshin.tests/META-INF/MANIFEST.MF
@@ -17,7 +17,8 @@ Require-Bundle: org.eclipse.core.resources;bundle-version="3.6.0",
  org.eclipse.emf.ecore;bundle-version="2.6.0";visibility:=reexport,
  org.eclipse.emf.ecore.xmi;bundle-version="2.6.0";visibility:=reexport,
  org.eclipse.emf.henshin.model;bundle-version="0.9.0",
- org.eclipse.emf.henshin.interpreter;bundle-version="0.9.0",
+ org.eclipse.emf.henshin.interpreter;bundle-version="1.9.0",
+ org.eclipse.emf.henshin.interpreter.debug;bundle-version="1.9.0",
  org.eclipse.emf.henshin.examples;bundle-version="0.9.0",
  org.eclipse.emf.henshin.statespace;bundle-version="0.9.0";visibility:=reexport,
  org.eclipse.emf.henshin.multicda.cpa;bundle-version="1.3.0",
