+++ plugins/org.eclipse.emf.henshin.tests/src/org/eclipse/emf/henshin/tests/basic/DebugTestBank.java
@@ -11,15 +11,16 @@ import org.eclipse.emf.henshin.examples.bank.BankExample;
 import org.eclipse.emf.henshin.interpreter.EGraph;
 import org.eclipse.emf.henshin.interpreter.Engine;
 import org.eclipse.emf.henshin.interpreter.Match;
-import org.eclipse.emf.henshin.interpreter.impl.DebugEngineImpl;
+import org.eclipse.emf.henshin.interpreter.impl.debug.DebugEngineImpl;
 import org.eclipse.emf.henshin.interpreter.impl.EGraphImpl;
 import org.eclipse.emf.henshin.interpreter.impl.MatchImpl;
 import org.eclipse.emf.henshin.interpreter.impl.UnitApplicationImpl;
 import org.eclipse.emf.henshin.interpreter.info.RuleInfo;
-import org.eclipse.emf.henshin.interpreter.matching.conditions.DebugApplicationCondition;
-import org.eclipse.emf.henshin.interpreter.matching.conditions.DebugApplicationCondition.ConstraintType;
-import org.eclipse.emf.henshin.interpreter.matching.conditions.DebugApplicationCondition.DebugLevel;
-import org.eclipse.emf.henshin.interpreter.matching.conditions.DebugApplicationCondition.DebugState;
+import org.eclipse.emf.henshin.interpreter.matching.conditions.ApplicationCondition;
+import org.eclipse.emf.henshin.interpreter.matching.conditions.debug.DebugApplicationCondition;
+import org.eclipse.emf.henshin.interpreter.matching.conditions.debug.DebugApplicationCondition.ConstraintType;
+import org.eclipse.emf.henshin.interpreter.matching.conditions.debug.DebugApplicationCondition.DebugLevel;
+import org.eclipse.emf.henshin.interpreter.matching.conditions.debug.DebugApplicationCondition.DebugState;
 import org.eclipse.emf.henshin.interpreter.matching.constraints.DomainSlot;
 import org.eclipse.emf.henshin.interpreter.matching.constraints.Variable;
 import org.eclipse.emf.henshin.model.Module;
