+++ plugins/org.eclipse.emf.henshin.tests/src/org/eclipse/emf/henshin/tests/basic/HenshinBreakpointsTest.java
@@ -20,15 +20,16 @@ import org.eclipse.emf.henshin.interpreter.Match;
 import org.eclipse.emf.henshin.interpreter.debug.DebugValueObject;
 import org.eclipse.emf.henshin.interpreter.debug.HenshinDebugValue;
 import org.eclipse.emf.henshin.interpreter.debug.HenshinDebugVariable;
-import org.eclipse.emf.henshin.interpreter.impl.DebugEngineImpl;
 import org.eclipse.emf.henshin.interpreter.impl.EGraphImpl;
 import org.eclipse.emf.henshin.interpreter.impl.MatchImpl;
 import org.eclipse.emf.henshin.interpreter.impl.UnitApplicationImpl;
+import org.eclipse.emf.henshin.interpreter.impl.debug.DebugEngineImpl;
 import org.eclipse.emf.henshin.interpreter.info.RuleInfo;
-import org.eclipse.emf.henshin.interpreter.matching.conditions.DebugApplicationCondition.ConstraintType;
-import org.eclipse.emf.henshin.interpreter.matching.conditions.DebugApplicationCondition.DebugLevel;
-import org.eclipse.emf.henshin.interpreter.matching.conditions.DebugApplicationCondition.DebugState;
-import org.eclipse.emf.henshin.interpreter.matching.conditions.TestDebugApplicationCondition;
+import org.eclipse.emf.henshin.interpreter.matching.conditions.ApplicationCondition;
+import org.eclipse.emf.henshin.interpreter.matching.conditions.debug.DebugApplicationCondition.ConstraintType;
+import org.eclipse.emf.henshin.interpreter.matching.conditions.debug.DebugApplicationCondition.DebugLevel;
+import org.eclipse.emf.henshin.interpreter.matching.conditions.debug.DebugApplicationCondition.DebugState;
+import org.eclipse.emf.henshin.interpreter.matching.conditions.debug.TestDebugApplicationCondition;
 import org.eclipse.emf.henshin.interpreter.matching.constraints.DomainSlot;
 import org.eclipse.emf.henshin.interpreter.matching.constraints.Variable;
 import org.eclipse.emf.henshin.model.Module;
